Add Configuration struct
Add new strcut for encapsulating the configuration options. Update
factory methods to accept the configuration struct to determine which
implementation of Reader and Writer to create. Update the HangmanGame
struct to include the configuration struct and use the configuration to
drive the difficulty of the word which is retrieved from the
WordGenerator.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>